그룹 컴포넌트 분리
그룹 탭에서 그룹 묶음들을 별도의 컴포넌트로 분리 (GroupContents)
